requirements.txt
ffmpeg-python
click
python-dotenv
pandas
moviepy
pytest
Pillow
pytesseract
opencv-python
rich
tqdm 